Allow short circuiting thrift network stack
This is useful in testing `Thrifter::Client` subclasses. This negates
the requirement for having to fake the transport and stub new to return
a mock/stub. This also has the nice side effect that URI does not need
to be set in the test environment if there is no network traffic.
Previously config.uri must have been set otherwise clients could not be
instantiated.