Add keep_alive config
This commit introduces to the "keep alive" config setting. When this
flag is set, the transport is only closed on error. Otherwise it's kept
open. The flag defaults to true. This changes the default behavior of
opening a new connection for every RPC. It may be disabled by setting
config.keep_alive = false.